backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Google Gemini AI
google-generativeai==0.8.3

# Image Processing
opencv-python==4.10.0.84
Pillow==11.0.0

# Environment Configuration
python-dotenv==1.0.1

# WebSocket Support
websockets==13.1